requirements.txt
pathlib==1.0.1
argparse==1.4.0
python-magic==0.4.27
Pillow==11.0.0
ExifRead==3.0.0